requirements.txt
openai
GitPython
prompt_toolkit

